# doc/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AriesHub API",
    "version": "0.7.0",
    "description": "Public content catalog, Redis-backed cookie identity, flat paged discussions with comment likes, administrator content and moderation API. Credit balance lives on users and pricing on publications; the credit ledger and unlock tables exist but the credit application API is not implemented yet. Private S3-compatible file uploads and owner-only temporary downloads. All JSON responses use Result {code,msg,data,traceId}; success code is SUCCESS, failures retain semantic HTTP status and business codes. Empty successes use HTTP 200 with data:null; throttling includes Retry-After and data.retryAfterSeconds."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1"
    }
  ],
  "paths": {
...
    "/auth/email-codes": {
      "post": {
        "operationId": "sendEmailCode",
        "summary": "Sends a six-digit registration code through Resend",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailCodeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Registration code request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/EmailCodeDispatch"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email or purpose",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "429": {
            "description": "Code resend cooldown or rate limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "503": {
            "description": "Email delivery failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "Creates a member account and signs it in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created; session cookie is set",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CurrentUser"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or verification code",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "429": {
            "description": "Registration rate limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "Signs in with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed in; session cookie is set",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/CurrentUser"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "429": {
            "description": "Login rate limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "Invalidates the current session",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success with an empty result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "code",
                    "msg",
                    "data",
                    "traceId"
                  ],
                  "properties": {
                    "code": {
                      "type": "string",
                      "const": "SUCCESS"
                    },
                    "msg": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    },
                    "traceId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Server-generated request trace ID; matches Result.traceId",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailCodeRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "purpose"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "purpose": {
            "type": "string",
            "enum": [
              "REGISTER"
            ]
          }
        }
      },
      "EmailCodeDispatch": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "expiresInSeconds",
          "resendAfterSeconds"
        ],
        "properties": {
          "expiresInSeconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "resendAfterSeconds": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password",
          "nickname",
          "code"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "maxLength": 72,
            "description": "Must contain at least one ASCII letter and one digit."
          },
          "nickname": {
            "type": "string",
            "minLength": 2,
            "maxLength": 30
          },
          "code": {
            "type": "string",
            "pattern": "^[0-9]{6}$"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "format": "password",
            "maxLength": 72
          }
        }
      },
...
      "CurrentUser": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "email",
          "nickname",
          "role",
          "emailVerified",
          "createdAt",
          "bio",
          "avatarFileId"
        ],
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[0-9]+$",
            "description": "PostgreSQL bigint ID serialized as a string."
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "nickname": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "bio": {
            "type": "string",
            "maxLength": 160
          },
          "avatarFileId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "msg",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "retryAfterSeconds"
                ],
                "properties": {
                  "retryAfterSeconds": {
                    "type": "integer",
                    "minimum": 1
                  }
                }
              }
            ]
          },
          "traceId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arieshub_token"
      }
    }
  }
}